Updates file names and references; chanfges loading text to a variable
