+++ src/components/bs5/loadingSpinner/LoadingSpinner.js
@@ -1,0 +1,11 @@
+import Component from "../../../js/QGDSComponent.js";
+import template from "./loadingSpinner.hbs?raw";
+
+export class LoadingSpinner {
+  // Use the global Component class to create a new instance of the Loading Spinner component.
+  // A data object, containing the Handlebars placeholder replacement strings, should be provided as an argument.
+
+  constructor(data = {}) {
+    return new Component(template, data);
+  }
+}
