+++ src/components/bs5/loadingSpinner/loadingSpinner.data.json
@@ -1,0 +1,5 @@
+{
+	"stacked": false,
+	"minimal": false,
+  "loadingText": "Loading..."
+}
