+++ src/components/bs5/loadingSpinner/loadingSpinner.hbs
@@ -1,8 +1,7 @@
 <!-- QGDS Component: Loading Spinner -->
-
 <div class="qld-spinner {{#if stacked}}align-center{{/if}}" role="status">
     <div class="spinner-border"></div>
     <span class="qld-spinner-label {{#if minimal}}sr-only{{/if}}">
-        Loading...
+        {{{ loadingText }}}
     </span>
 </div>
