--- src/components/bs5/spinner/spinner.data.json
@@ -1,4 +1,0 @@
-{
-	"stacked": false,
-	"minimal": false
-}
