+++ src/css/main.scss
@@ -102,7 +102,7 @@ $enable-dark-mode: true;
 @import "../components/bs5/quickexit/quickexit";
 @import "../components/bs5/sidenav/sidenav";
 @import "../components/bs5/searchInput/searchInput";
-@import "../components/bs5/spinner/spinner";
+@import "../components/bs5/loadingSpinner/loadingSpinner";
 @import "../components/bs5/table/table";
 @import "../components/bs5/tag/tag";
 @import "../components/bs5/textbox/textInput";
