+++ src/js/handlebars.partials.js
@@ -26,6 +26,7 @@ import head from "../components/bs5/head/head.hbs?raw";
 import header from "../components/bs5/header/header.hbs?raw";
 import inpageAlert from "../components/bs5/inpageAlert/inpageAlert.hbs?raw";
 import inpagenav from "../components/bs5/inpagenav/inpagenav.hbs?raw";
+import loadingSpinner from "../components/bs5/loadingSpinner/loadingSpinner.hbs?raw";
 import mainContainerWrapper from "../components/bs5/mainContainerWrapper/mainContainerWrapper.hbs?raw";
 import metaDcTerms from "../components/bs5/metaDcTerms/metaDcTerms.hbs?raw";
 import metaOpenGraph from "../components/bs5/metaOpenGraph/metaOpenGraph.hbs?raw";
@@ -37,7 +38,6 @@ import searchInput from "../components/bs5/searchInput/searchInput.hbs?raw";
 import select from "../components/bs5/select/select.hbs?raw";
 import sidenav from "../components/bs5/sidenav/sidenav.hbs?raw";
 import sidenavWrapper from "../components/bs5/sidenavWrapper/sidenavWrapper.hbs?raw";
-import spinner from "../components/bs5/spinner/spinner.hbs?raw";
 import table from "../components/bs5/table/table.hbs?raw";
 import tag from "../components/bs5/tag/tag.hbs?raw";
 import textarea from "../components/bs5/textarea/textarea.hbs?raw";
@@ -77,6 +77,7 @@ export default function handlebarsPartials(handlebars) {
   handlebars.registerPartial("header", header);
   handlebars.registerPartial("inpageAlert", inpageAlert);
   handlebars.registerPartial("inpagenav", inpagenav);
+  handlebars.registerPartial("loadingSpinner", loadingSpinner);
   handlebars.registerPartial("mainContainerWrapper", mainContainerWrapper);
   handlebars.registerPartial("metaDcTerms", metaDcTerms);
   handlebars.registerPartial("metaOpenGraph", metaOpenGraph);
@@ -88,7 +89,6 @@ export default function handlebarsPartials(handlebars) {
   handlebars.registerPartial("select", select);
   handlebars.registerPartial("sidenav", sidenav);
   handlebars.registerPartial("sidenavWrapper", sidenavWrapper);
-  handlebars.registerPartial("spinner", spinner);
   handlebars.registerPartial("table", table);
   handlebars.registerPartial("tag", tag);
   handlebars.registerPartial("textarea", textarea);
